requirements.txt
beautifulsoup4
requests
python-dotenv
pytest
selenium
pyvirtualdisplay